+++ src/main/resources/db/migration/V1__init.sql
@@ -2,8 +2,8 @@ CREATE TYPE role AS ENUM ('OWNER', 'ADMIN', 'USER');
 
 create table account
 (
-    id integer not null primary key unique,
-    chat_id integer not null,
+    id bigint not null primary key unique,
+    chat_id bigint not null,
     username VARCHAR(128),
     role role not null
 );
